==> 1 GetAllTriples.ipynb <==
import pandas as pd
from os.path import join
from glob import glob
from tqdm.notebook import tqdm
import gc

def collect_garbage():
    print(gc.get_count())
    gc.collect()
    print(gc.get_count())
# my_dtypes = OrderedDict([("time",int), ("cik",int)])
PREDICATION_COLS = [
    'PREDICATION_ID', 'SENTENCE_ID', 'PMID', 'PREDICATE', 
    'SUBJECT_CUI', 'SUBJECT_NAME', 'SUBJECT_SEMTYPE', 'SUBJECT_NOVELTY', 'OBJECT_CUI',
    'OBJECT_NAME', 'OBJECT_SEMTYPE', 'OBJECT_NOVELTY', 'Nan1', 'Nan2', 'Nan3']
TRIPLE_DATA = ['PREDICATE','SUBJECT_CUI','SUBJECT_NAME',
               'SUBJECT_SEMTYPE','OBJECT_CUI','OBJECT_NAME','OBJECT_SEMTYPE']
PREDICATION_AUX_COLS = ['PREDICATION_AUX_ID' ,'PREDICATION_ID' ,'SUBJECT_TEXT','SUBJECT_DIST'
,'SUBJECT_MAXDIST','SUBJECT_START_INDEX'
,'SUBJECT_END_INDEX','SUBJECT_SCORE'
,'INDICATOR_TYPE','PREDICATE_START_INDEX'
,'PREDICATE_END_INDEX','OBJECT_TEXT','OBJECT_DIST'
,'OBJECT_MAXDIST','OBJECT_START_INDEX'
,'OBJECT_END_INDEX','OBJECT_SCORE'
,'CURR_TIMESTAMP']

PREDICATION_DIR = 'PREDICATION'
PREDICATION_AUX_DIR = 'PREDICATION_AUX'
TRIPLES_DIR = 'all_triples'

collect_garbage()

## solve the cui canonization problem
## create link between canonized triples and original triples
## create link between original triples and predications


def canonize_cui(cui):
    cacon_form = 'C0000000'
    if len(cui)<8:
        cui = cacon_form[0:-len(cui)]+cui
    elif 'C' not in  cui:
        cui = 'C' + cui
    return cui

def process_mult_cui(cui):
    origin_cui = None
    all_cuis = []
    for cu in cui.split('|'):
        if origin_cui is None:
            cu = canonize_cui(cu)
            origin_cui = cu
        else:
#             cu = origin_cui[0:len(origin_cui)-len(cu)] + cu
            cu = origin_cui[0:-len(cu)] + cu
        all_cuis.append(cu)
    return all_cuis

def process_cui(cui):
    if '|' in cui:
        cuis = process_mult_cui(cui)
    else:
        cuis = [canonize_cui(cui)]
    return cuis

def get_all_triples():
    all_files = sorted(glob(join(PREDICATION_DIR,'*.gz')))
    triple_df = None
    entity_df = None
    for f_name in tqdm(all_files):
        print('--------')
        df = pd.read_csv(f_name,  compression= 'gzip', sep = ',', encoding='iso-8859-1', header=None, names = PREDICATION_COLS)[TRIPLE_DATA]
        print('origin df .shape: ',df.shape)
        sub_data = df[['SUBJECT_CUI','SUBJECT_SEMTYPE', 'SUBJECT_NAME']].rename(columns ={'SUBJECT_CUI':'ENTITY','SUBJECT_SEMTYPE':'SEMTYPE', 'SUBJECT_NAME':'NAME'})
        obj_data = df[['OBJECT_CUI','OBJECT_SEMTYPE', 'OBJECT_NAME']].rename(columns ={'OBJECT_CUI':'ENTITY','OBJECT_SEMTYPE': 'SEMTYPE', 'OBJECT_NAME':'NAME'})
#         print(sub_data)
        
#         print('obj_data .shape: ',obj_data.shape)
#         print('sub_data .shape: ',sub_data.shape)
        ent_data = pd.concat([sub_data, obj_data], axis=0, ignore_index=True)
#         print(ent_data)
        ent_data.drop_duplicates(inplace=True, ignore_index=True)
        df.drop(['SUBJECT_NAME','SUBJECT_SEMTYPE','OBJECT_NAME','OBJECT_SEMTYPE'], inplace=True, axis=1)
        df.drop_duplicates(inplace=True, ignore_index=True)
#         print('ent_data.shape: ',ent_data.shape)
#         print('dropped df .shape: ',df.shape)
        
        if entity_df is None:
            entity_df = ent_data
        else:
            entity_df = pd.concat([entity_df, ent_data], axis=0, ignore_index=True)
            entity_df.drop_duplicates(inplace=True, ignore_index=True)
#         if len(list(entity_df.columns))>2:
#             print(ent_data)
        print('entity_df.shape: ',entity_df.shape)
        if triple_df is None:
            triple_df = df
        else:
            triple_df = pd.concat([triple_df, df], axis=0, ignore_index=True)
            triple_df.drop_duplicates(inplace=True, ignore_index=True)
        print('triple_df.shape: ',triple_df.shape)
        

    return triple_df, entity_df

def get_processed_triples_entities():
    pass

all_triples = get_all_triples()

all_triples[0]

pd.read_csv(join(TRIPLES_DIR, 'all_data_triples.csv'), compression = 'gzip')

all_triples[0].to_csv(join(TRIPLES_DIR, 'all_data_triples.csv'), compression = 'gzip', index = False)

all_triples[1].to_csv(join(TRIPLES_DIR, 'all_data_entities.csv'), compression = 'gzip', index = False)

all_triples[1]

## Triple _ Predication map

def get_trip_2_id(trip_df):
    PREDICATE = list(trip_df['PREDICATE'])
    SUBJECT_CUI = list(trip_df['SUBJECT_CUI'])
    OBJECT_CUI = list(trip_df['OBJECT_CUI'])
    index = list(trip_df.index)
    return {(p, s, o):i for p, s, o, i in zip(PREDICATE, SUBJECT_CUI, OBJECT_CUI, index)}

def get_triple_pred_map(trip_df):
    trip_2_id = get_trip_2_id(trip_df)
    all_files = sorted(glob(join(PREDICATION_DIR,'*.gz')))
    all_dfs = []
    
    for f_name in tqdm(all_files):
#         print('--------')
#         df = pd.read_csv(f_name,  compression= 'gzip', sep = ',', encoding='iso-8859-1', header=None, names = PREDICATION_COLS, engine='pyarrow')
        df = pd.read_csv(
            f_name, compression= 'gzip', on_bad_lines=test_line, encoding='ISO-8859-1',
            header=None, names = PREDICATION_COLS, engine='python'
        )
        df = df.groupby(['PREDICATE','SUBJECT_CUI','OBJECT_CUI'])[['PREDICATION_ID', 'SENTENCE_ID']].agg(lambda x: list(x))
        df["ORIGIN_ID"] = [trip_2_id[a] for a in list(df.index)]
        df = df[['ORIGIN_ID', 'PREDICATION_ID', 'SENTENCE_ID']]
        all_dfs.append(df)
    all_dfs = pd.concat(all_dfs, ignore_index = True)
#     all_dfs[['ORIGIN_ID', 'SENTENCE_ID']].explode('SENTENCE_ID')
#     all_dfs[['ORIGIN_ID', 'PREDICATION_ID']].explode('PREDICATION_ID')
#     all_dfs = all_dfs.drop_duplicates(ignore_index = True)
    return all_dfs

def get_triple_pred_aux_map(trip_df):
    trip_2_id = get_trip_2_id(trip_df)
    all_files = sorted(glob(join(PREDICATION_AUX_DIR,'*.gz')))
    all_dfs = []
    
    for f_name in tqdm(all_files):
#         print('--------')
        df = pd.read_csv(f_name,  compression= 'gzip', sep = ',', encoding='iso-8859-1', header=None, names = PREDICATION_AUX_COLS, engine='pyarrow')
#         return df
        df = df.groupby(['PREDICATE','SUBJECT_CUI','OBJECT_CUI'])[['PREDICATION_AUX_ID', 'PREDICATION_ID']].agg(lambda x: list(x))
        df["ORIGIN_ID"] = [trip_2_id[a] for a in list(df.index)]
        df = df[['ORIGIN_ID', 'PREDICATION_ID', 'SENTENCE_ID']]
        all_dfs.append(df)
    all_dfs = pd.concat(all_dfs, ignore_index = True)
    return all_dfs
def test_line(bad_line):
    print(bad_line)
    return bad_line

def get_pred_aux_pred_map():
    all_files = sorted(glob(join(PREDICATION_AUX_DIR,'*.gz')))
    all_dfs = []
    
    for f_name in tqdm(all_files):
        df = pd.read_csv(
            f_name, compression= 'gzip', on_bad_lines=test_line, encoding='ISO-8859-1',
            header=None, names = PREDICATION_AUX_COLS, engine='python'
        )
#         df = pd.read_csv(f_name,  compression= 'gzip', sep = ',', encoding='iso-8859-1', header=None, names = PREDICATION_AUX_COLS, engine='pyarrow')
        all_dfs.append(df[['PREDICATION_AUX_ID', 'PREDICATION_ID']])
    all_dfs = pd.concat(all_dfs, ignore_index = True)
    return all_dfs

def get_triple_pred_map_df(res_dict):
    all_data = []
    for ORIGIN_ID in res_dict:
        data_ = {}
        data_['ORIGIN_ID'] = ORIGIN_ID
        for PREDICATION_ID in res_dict[ORIGIN_ID]:
            data_['PREDICATION_ID'] = PREDICATION_ID
            all_data.append(data_)
    return pd.DataFrame(all_data)
        

all_triples = pd.read_csv(join(TRIPLES_DIR, 'all_data_triples.csv'), compression = 'gzip')
all_entities = pd.read_csv(join(TRIPLES_DIR, 'all_data_entities.csv'), compression = 'gzip')

# xx = get_trip_2_id(all_triples)
del all_triples
del all_entities

pred_aux_pred_map = get_pred_aux_pred_map()

pred_aux_pred_map.to_csv(join('index_maps', 'pred_aux_pred_map.csv'), index = False, compression = 'gzip')

pred_aux_pred_map[pred_aux_pred_map['PREDICATION_AUX_ID']== 92150804]

#del xx
# id_2_pred_aux = get_triple_pred_aux_map(all_triples)

id_2_pred_aux

id_2_pred = get_triple_pred_map(all_triples)

id_2_pred

id_2_pred = id_2_pred.set_index(['ORIGIN_ID']).apply(pd.Series.explode).reset_index()

id_2_pred

id_2_pred.to_csv(join('index_maps','origin_predication.csv'), compression = 'gzip', index = False)

del id_2_pred

## Origin Pred AUX MAP

def get_pred_pred_aux_origin_sent_map():
    ori_pred = pd.read_csv(join('index_maps','origin_predication.csv'), compression = 'gzip')
    pred_aux_pred = pd.read_csv(join('index_maps', 'pred_aux_pred_map.csv'), compression = 'gzip')
    return ori_pred.merge(pred_aux_pred,how='inner', left_on='PREDICATION_ID', right_on='PREDICATION_ID')

#     pred_aux_pred = dict(zip(pred_aux_pred['PREDICATION_ID'], pred_aux_pred['PREDICATION_AUX_ID']))
#     ori_pred['PREDICATION_AUX_ID'] = ori_pred['PREDICATION_ID'].map(pred_aux_pred)
#     return ori_pred


all_map = get_pred_pred_aux_origin_sent_map()

all_map

all_map[all_map['PREDICATION_AUX_ID']==92150156]

all_map.set_index('PREDICATION_AUX_ID').loc[92150156, :]

all_map.to_csv(join('index_maps','origin_pred_pred_aux_sent_map.csv'), compression = 'gzip', index = False)

## Canonizing CUIs

def canonize_cui(cui):
    cacon_form = 'C0000000'
    if len(cui)<8:
        cui = cacon_form[0:-len(cui)]+cui
    elif 'C' not in  cui:
        cui = 'C' + cui
    return cui

def process_mult_cui(cui):
    origin_cui = None
    all_cuis = []
    for cu in cui.split('|'):
        if origin_cui is None:
            cu = canonize_cui(cu)
            origin_cui = cu
        else:
#             cu = origin_cui[0:len(origin_cui)-len(cu)] + cu
            if len(origin_cui)>=8:
                cu = origin_cui[0:-len(cu)] + cu
            else:
                cu = canonize_cui(cu)
                origin_cui = cu
        if len(origin_cui)>=8:
            all_cuis.append(cu)
    return all_cuis

def process_cui(cui):
    if '|' in cui:
        cuis = process_mult_cui(cui)
    else:
        cuis = [canonize_cui(cui)]
    return cuis

def cononize_entities(ent_df):
    all_data = []
    for i, rec in ent_df.iterrows():
        ent_data = {}
        ENTITY = rec['ENTITY']
        ent_data['SEMTYPE'] = rec['SEMTYPE']
        ent_data['NAME']  = rec['NAME']
        ent_data['ORIGIN_ID'] = i
        cuis = process_cui(ENTITY)
        for cui in cuis:
            ent_data['ENTITY'] = cui
            all_data.append(ent_data)
    return pd.DataFrame(all_data)

def canonize_triples(trip_df):
    all_data = []
    for i, rec in trip_df.iterrows():
        trip_data = {}
        trip_data['PREDICATE'] = rec['PREDICATE']
        trip_data['ORIGIN_ID'] = i
        
        SUBJECT_CUI = rec['SUBJECT_CUI']
        OBJECT_CUI = rec['OBJECT_CUI']
        
        sub_cuis = process_cui(SUBJECT_CUI)
        obj_cuis = process_cui(OBJECT_CUI)
        for sub_cui in sub_cuis:
            for obj_cui in obj_cuis:
                trip_data['SUBJECT_CUI'] = sub_cui
                trip_data['OBJECT_CUI'] = obj_cui
                all_data.append(trip_data)
    return pd.DataFrame(all_data)

all_triples = pd.read_csv(join(TRIPLES_DIR, 'all_data_triples.csv'), compression = 'gzip')
all_entities = pd.read_csv(join(TRIPLES_DIR, 'all_data_entities.csv'), compression = 'gzip')

can = cononize_entities(all_entities)

#all_entities.shape

#can

#all_entities[all_entities['NAME']=='1|podg|humn']

#all_entities[all_entities['NAME']=='1|podg|humn']

#can[can['ENTITY']=='Patients']

#all_entities[all_entities['ENTITY']=='Patients']

#{i for i in list(can['ENTITY'])if 'C' not in i and len(i)<=8}

#{i for i in list(can['ENTITY'])if i==''}

#process_mult_cui('10128')

#{i for i in list(all_entities['ENTITY'])if '10128' in i}

#{i for i in list(can['ENTITY'])if '|' in i}

#canonize_cui('10128')

can.to_csv(join(TRIPLES_DIR, 'all_data_entities_can.csv'), compression = 'gzip', index = False)

can = canonize_triples(all_triples)

can

#process_mult_cui('10128')

{i for i in list(can['SUBJECT_CUI'])if 'C' not in i and len(i)<=8}

{i for i in list(can['SUBJECT_CUI'])if '|' in i}

can.to_csv(join(TRIPLES_DIR, 'all_data_triples_can.csv'), compression = 'gzip', index = False)

del can

## Tests


df = pd.read_csv('PREDICATION/split_07.csv.gz',  compression= 'gzip', sep = ',', engine="pyarrow",header=None, names = PREDICATION_COLS)
%time

df = pd.read_csv('PREDICATION/split_26.csv.gz', compression= 'gzip', encoding='iso-8859-1', sep = ',',header=None, names = PREDICATION_COLS)
%time

df

pd.io.parquet.get_engine('auto')

pd.io.parquet.PyArrowImpl()

pd.show_versions()

df

sorted(glob(join(PREDICATION_DIR,'*.gz')))[6]
