==> 3 Splite Sentence Files and AUX_Predications.ipynb <==
# the objective is to splite sentence files and aux_predications into multiple files
# each file needs to contains only sentences and aux_preds of the same triple
import pandas as pd
from glob import glob
from os.path import join
import gc
import gzip
from csv import reader
from collections import Counter
import os
from tqdm.notebook import tqdm

PREDICATION_COLS = [
    'PREDICATION_ID', 'SENTENCE_ID', 'PMID', 'PREDICATE', 
    'SUBJECT_CUI', 'SUBJECT_NAME', 'SUBJECT_SEMTYPE', 'SUBJECT_NOVELTY', 'OBJECT_CUI',
    'OBJECT_NAME', 'OBJECT_SEMTYPE', 'OBJECT_NOVELTY', 'Nan1', 'Nan2', 'Nan3']
SENTENCE_COLS = ['SENTENCE_ID','PMID','TYPE','NUMBER','SENT_START_INDEX','SENTENCE',
                 'SENT_END_INDEX','NORMALIZED_SECTION_HEADER','SECTION_HEADER']
CITATIONS_COLS = ['PMID','ISSN','DP','EDAT','PYEAR']

PREDICATION_DIR = 'PREDICATION'
SENTENCE_DIR = 'SENTENCE'

TRIPLES_DIR = 'all_triples'
TRIP_SENT_DIR = 'TRIPLE_SENTENCES'
CITATIONS_DIR = 'CITATIONS'

def collect_garbage():
    print(gc.get_count())
    gc.collect()
    print(gc.get_count())


collect_garbage()

or_preds = pd.read_csv(join(TRIPLES_DIR,'origin_predication.csv'), compression = 'gzip')

or_preds.shape

or_preds.drop_duplicates().shape

def manual_separation(bad_line):
    good_1 = bad_line[:5]
    good_2 = bad_line[-3:]
    sent_parts = bad_line[5:-3]
    sent_parts = [', '.join(sent_parts)]
    all_data = good_1+ sent_parts + good_2
    return all_data

def get_triple_sentences(groups_save):
    all_files = sorted(glob(join(SENTENCE_DIR,'*.gz')))
    print(len(all_files))
    for fil in tqdm(all_files):
        df = pd.read_csv(fil, on_bad_lines=manual_separation,
                 compression= 'gzip', encoding='ISO-8859-1', header=None, names = SENTENCE_COLS, engine='python'
                        )
        df = pd.merge(or_preds,df,on='SENTENCE_ID')
        for name, group in df.groupby('ORIGIN_ID'):
            group_dir = str(name)
            if not os.path.exists(join(groups_save,group_dir)):
                os.mkdir(join(groups_save,group_dir))
            num_files = len(glob(join(groups_save,group_dir,'*')))
            save_file = str(num_files + 1)+'.csv'
#             print('*'*15,name)
#             print('num_files:',num_files)
#             print(group)
            
            group.to_csv(join(groups_save,group_dir,save_file), compression = 'gzip')

def map_sentences_2_files():
    all_files = sorted(glob(join(SENTENCE_DIR,'*.gz')))
    all_res = []
    for fil in tqdm(all_files):
        f_name = fil.split('/')[-1].split('.')[0]
        df = pd.read_csv(fil, on_bad_lines=manual_separation,
                 compression= 'gzip', encoding='ISO-8859-1', header=None, names = SENTENCE_COLS, engine='python'
        )
        df['f_name'] = f_name
        all_res.append(df[['f_name', 'SENTENCE_ID']])
    return pd.concat(all_res, ignore_index=False)

sent_file_map = map_sentences_2_files()

sent_file_map.to_csv(join(TRIPLES_DIR, 'sentence_index.csv'), index = False, compression = 'gzip')

#test_ = get_triple_sentences(TRIP_SENT_DIR)

test_

## Triple Time Validity

all_files = glob(join(TRIP_SENT_DIR, '*', '*'))

all_files = glob(join('TRIPLE_SENTENCES/1805287', '*'))

pd.read_csv(all_files[0], compression = 'gzip')

or_pred = pd.read_csv(join(TRIPLES_DIR, 'origin_predication.csv'), compression = 'gzip')

or_pred.head()

citations = pd.read_csv(join(CITATIONS_DIR, 'semmedVER43_2021_R_CITATIONS.csv'), compression = 'gzip',
                        encoding='ISO-8859-1', header = None, names = CITATIONS_COLS)

citations.head()

citations['EDAT'].str.split('-')
