==> 4-5 Adding Time to the UKG.ipynb <==
import pandas as pd
from os.path import join
from glob import glob
from tqdm.notebook import tqdm
from multiprocessing import Pool # for reading the CSVs faster

from multiprocessing.pool import ThreadPool
import json

TRIPLES_DIR = 'all_triples'
CITATIONS_DIR = 'CITATIONS'
TRIP_SENT_DIR = 'TRIPLE_SENTENCES'
CITATIONS_COLS = ['PMID','ISSN','DP','EDAT','PYEAR']


proba_df = pd.read_csv(join('all_triples', 'triples_probabilities.csv'), compression = 'gzip')

proba_df.head()

def process_citations():
    cit_df = pd.read_csv(join(CITATIONS_DIR, 'semmedVER43_2021_R_CITATIONS.csv'), compression = 'gzip',
                        encoding='ISO-8859-1', header = None, names = CITATIONS_COLS)
    cit_df_ = cit_df[~cit_df['EDAT'].isna()]
    cit_df_['EDAT'] = pd.to_datetime(cit_df_['EDAT'])
    cit_df_no = cit_df[cit_df['EDAT'].isna()]
    cit_df_no['EDAT'] = pd.to_datetime(cit_df_no['DP'])
    return pd.concat([cit_df_, cit_df_no])

clean_cite = process_citations()

clean_cite.shape

clean_cite.head()

cite_date_map = dict(zip(clean_cite['PMID'], clean_cite['EDAT']))

cite_date_map

clean_cite['EDAT'].min()

clean_cite['EDAT'].max()

def write_json_lines(file_name,dict_data):
    json_string = json.dumps(dict_data)
    with open(file_name, 'a') as f:
        f.write(json_string+"\n")
        
def read_json_lines(file_name):
    lines = []
    with open(file_name) as file_in:
        for line in file_in:
            lines.append(json.loads(line))
    return lines

def my_read_csv(filename):
    # Helper function for the parellel load_csvs
    return pd.read_csv(filename, compression = 'gzip')

def load_csvs(parent_path):
    """Reads and joins all our CSV files into one big dataframe.
    We do it in parallel to make it faster, since otherwise it takes some time.
    Idea from: https://stackoverflow.com/questions/36587211/easiest-way-to-read-csv-files-with-multiprocessing-in-pandas
    
    """
    # set up your pool
    pool = ThreadPool(10) 
    file_list = glob(join(parent_path, '*'))
    df_list = pool.map(my_read_csv, file_list)
    if len(df_list)>0:
        return pd.concat(df_list, ignore_index=True)
    else:
        return None

def get_all_csvs(parent_path):
    all_dfs = []
    for i in glob(join(parent_path, '*')):
        df_ = pd.read_csv(i, compression = 'gzip')
        all_dfs.append(df_)
    if len(all_dfs)>0:
        return pd.concat(all_dfs)
    else:
        return None
    
def get_el(search_dict):
    def get_el_(i):
        return search_dict[i]
    return get_el_
def time_validity_triples(p_df, pmid_time_map, save_file):
    res = []
    for ix, ORIGIN_ID in enumerate(tqdm(list(set(p_df['ORIGIN_ID'])))):
        res_ = {}
#         all_df = get_all_csvs(join(TRIP_SENT_DIR, str(ORIGIN_ID)))
        all_df = load_csvs(join(TRIP_SENT_DIR, str(ORIGIN_ID)))
        res_['ORIGIN_ID'] = ORIGIN_ID
        if all_df is not None:
            mapped_time = all_df['PMID'].apply(get_el(pmid_time_map))
            res_['time_min'] = str(mapped_time.min())
            res_['time_max'] = str(mapped_time.max())
            res_['true_time'] = True
        else:
            res_['time_min'] = ''
            res_['time_max'] = ''
            res_['true_time'] = False
#         res.append(res_)
        write_json_lines(file_name = save_file,dict_data = res_)
#     return pd.DataFrame(res)

origin_time = time_validity_triples(p_df = proba_df,
                                    pmid_time_map = cite_date_map, save_file=join(TRIPLES_DIR, 'origin_time.jsonl'))

str(origin_time)
