==> 4-7 Triple Spliting.ipynb <==
import pandas as pd
from os.path import join
import json
import numpy as np

PROC_UMLS = '/home/pc/Desktop/AdilStuff/Projects/SemRepMed/umls processed net data'
SELECTED_TRIPLES = 'Selected Triples'
# COVID Complet  | risk factor covid kg
# SELECTED_KG = 'risk factor covid kg'
SELECTED_KG = 'risk_factor_disease_time'
TRIPLES_DIR = 'all_triples'

def save_json(path_, data_):
    with open(path_, 'w') as f:
        json.dump(data_, f)
        
def read_json(path_):
    with open(path_, 'r') as f:
        data = json.load(f)
    return data

def verify_splits(train_, test_, val_,
                  head_col = 'SUBJECT_CUI', rel_col = 'PREDICATE', tail_col = 'OBJECT_CUI',
                  proba_col = 'label_proba'):
    
    all_train_ents = set(list(train_[head_col]) + list(train_[tail_col]))
    all_train_rels = set(list(train_[rel_col]))
    total_triples = train_.shape[0]
    all_probas = list(train_[proba_col])
    if val_ is not None:
        all_val_ents = set(list(val_[head_col]) + list(val_[tail_col]))
        all_val_rels = set(list(val_[rel_col]))
        total_triples += val_.shape[0]
        all_probas += list(val_[proba_col])
    else:
        all_val_ents, all_val_rels = [], []
        
    if test_ is not None:
        all_test_ents = set(list(test_[head_col]) + list(test_[tail_col]))
        all_test_rels = set(list(test_[rel_col]))
        total_triples += test_.shape[0]
        all_probas += list(test_[proba_col])
    else:
        all_test_ents, all_test_rels = [], []
        
    print('total number of triples: ',total_triples)
    print('# all entities: ',len(set(list(all_val_ents)+list(all_test_rels)+list(all_train_ents))))
    print('# all relations: ',len(set(list(all_test_rels)+list(all_val_rels)+list(all_train_rels))))
    print('# Average proba: ', np.array(all_probas).mean())
    print('# Std proba: ', np.array(all_probas).std())
    
    print('number of training triples: ',train_.shape[0])
    if test_ is not None:
        print('number of test triples: ',test_.shape[0])
    if val_ is not None:
        print('number of val triples: ',val_.shape[0])
    print('fraction of training triples: ',train_.shape[0]/total_triples)
    if test_ is not None:
        print('fraction of test triples: ',test_.shape[0]/total_triples)
        
    if val_ is not None:
        print('fraction of val triples: ',val_.shape[0]/total_triples)
    
    if (train_ is not None) & (test_ is not None) & (val_ is not None):
        print('ents in val not train: ',len(all_val_ents.difference(all_train_ents)))
        print('ents in test not train: ',len(all_test_ents.difference(all_train_ents)))
        print('rels in val not train: ',len(all_val_rels.difference(all_train_rels)))
        print('rels in test not train: ',len(all_test_rels.difference(all_train_rels)))
    
def kg_split():
    all_df = pd.read_csv(join(SELECTED_TRIPLES, SELECTED_KG, 'full_kg.csv'), compression = 'gzip')
    train_df = all_df.groupby(['SUBJECT_CUI', 'OBJECT_CUI']).sample(frac = 0.56)
    test_val = all_df[~all_df.index.isin(train_df.index)]
    test_df = test_val.sample(frac = 0.5)
    val_df = test_val[~test_val.index.isin(test_df.index)]
    return train_df, test_df, val_df

def splits_formatting_save(train_, test_, val_, save_path):
    
    new_col_names = {'PREDICATE':'r','SUBJECT_CUI':'h','OBJECT_CUI':'t',
                     'label_proba':'p','SUBJECT_TYPE':'h_type','OBJECT_TYPE':'t_type'}
    train_.rename(new_col_names, axis = 1).to_csv(join(save_path, 'raw_train.csv'), index = False)
    test_.rename(new_col_names, axis = 1).to_csv(join(save_path, 'raw_test.csv'), index = False)
    val_.rename(new_col_names, axis = 1).to_csv(join(save_path, 'raw_val.csv'), index = False)
    all_train_ents = list(train_['SUBJECT_CUI']) + list(train_['OBJECT_CUI'])
    all_train_ents_types = list(train_['SUBJECT_TYPE']) + list(train_['OBJECT_TYPE'])
    all_train_rels = list(train_['PREDICATE'])
    
    all_val_ents = list(val_['SUBJECT_CUI']) + list(val_['OBJECT_CUI'])
    all_val_ents_types = list(val_['SUBJECT_TYPE']) + list(val_['OBJECT_TYPE'])
    all_val_rels = list(val_['PREDICATE'])
    
    all_test_ents = list(test_['SUBJECT_CUI']) + list(test_['OBJECT_CUI'])
    all_test_ents_types = list(test_['SUBJECT_TYPE']) + list(test_['OBJECT_TYPE'])
    all_test_rels = list(test_['PREDICATE'])
    
    all_ents = list(set(all_train_ents + all_val_ents + all_test_ents))
    all_rels = list(set(all_train_rels + all_val_rels + all_test_rels))
    all_ent_types = list(set(all_train_ents_types + all_val_ents_types + all_test_ents_types))
    ent_map = {e:idx for idx, e in enumerate(all_ents)}
    rel_map = {e:idx for idx, e in enumerate(all_rels)}
    ent_type_map = {e:idx for idx, e in enumerate(all_ent_types)}
    #####
    train_['SUBJECT_CUI'] = train_['SUBJECT_CUI'].map(ent_map)
    train_['OBJECT_CUI'] = train_['OBJECT_CUI'].map(ent_map)
    train_['SUBJECT_TYPE'] = train_['SUBJECT_TYPE'].map(ent_type_map)
    train_['OBJECT_TYPE'] = train_['OBJECT_TYPE'].map(ent_type_map)
    train_['PREDICATE'] = train_['PREDICATE'].map(rel_map)
    #
    test_['SUBJECT_CUI'] = test_['SUBJECT_CUI'].map(ent_map)
    test_['OBJECT_CUI'] = test_['OBJECT_CUI'].map(ent_map)
    test_['SUBJECT_TYPE'] = test_['SUBJECT_TYPE'].map(ent_type_map)
    test_['OBJECT_TYPE'] = test_['OBJECT_TYPE'].map(ent_type_map)
    test_['PREDICATE'] = test_['PREDICATE'].map(rel_map)
    #
    val_['SUBJECT_CUI'] = val_['SUBJECT_CUI'].map(ent_map)
    val_['OBJECT_CUI'] = val_['OBJECT_CUI'].map(ent_map)
    val_['SUBJECT_TYPE'] = val_['SUBJECT_TYPE'].map(ent_type_map)
    val_['OBJECT_TYPE'] = val_['OBJECT_TYPE'].map(ent_type_map)
    val_['PREDICATE'] = val_['PREDICATE'].map(rel_map)
    ####
    train_ = train_.rename(new_col_names, axis = 1)
    test_ = test_.rename(new_col_names, axis = 1)
    val_ = val_.rename(new_col_names, axis = 1)
    ## saving
    train_.to_csv(join(save_path, 'train.csv'), index = False)
    test_.to_csv(join(save_path, 'test.csv'), index = False)
    val_.to_csv(join(save_path, 'val.csv'), index = False)
    save_json(join(save_path, 'ent_map.json'), ent_map)
    save_json(join(save_path, 'rel_map.json'), rel_map)
    save_json(join(save_path, 'ent_type_map.json'), ent_type_map)
    return train_, test_, val_, ent_map, rel_map, ent_type_map

def test_read(save_path):
    train = pd.read_csv(join(save_path, 'train.csv'))
    test = pd.read_csv(join(save_path, 'test.csv'))
    val = pd.read_csv(join(save_path, 'val.csv'))
    raw_train = pd.read_csv(join(save_path, 'raw_train.csv'))
    raw_test = pd.read_csv(join(save_path, 'raw_test.csv'))
    raw_val = pd.read_csv(join(save_path, 'raw_val.csv'))
    ent_map = read_json(join(save_path, 'ent_map.json'))
    rel_map = read_json(join(save_path, 'rel_map.json'))
    ent_type_map = read_json(join(save_path, 'ent_type_map.json'))
    return raw_train, raw_test, raw_val, train, test, val, ent_map, rel_map, ent_type_map

train_df, test_df, val_df = kg_split()

verify_splits(train_df, test_df, val_df)

train_df

save_path = join(SELECTED_TRIPLES, SELECTED_KG)


pd.read_csv(join(save_path, 'raw_train.csv'))

splits_formatting_save(train_ = train_df.copy(), test_ = test_df.copy(), val_ = val_df.copy(), save_path = save_path)

save_path = join(SELECTED_TRIPLES, SELECTED_KG)
raw_train, raw_test, raw_val, train, test, val, ent_map, rel_map, ent_type_map = test_read(save_path)

train

len(rel_map)

len(ent_type_map)

len(ent_map)

rel_map

## Get Stats

save_path = join(SELECTED_TRIPLES, SELECTED_KG)
raw_train, raw_test, raw_val, train, test, val, ent_map, rel_map, ent_type_map = test_read(save_path)

pd.read_csv(join(save_path, 'raw_train.csv'))

verify_splits(raw_train, raw_test, raw_val,
                  head_col = 'h', rel_col = 'r', tail_col = 't',
                  proba_col = 'p')

## Total KG

proba_triples_df = pd.read_csv(join(TRIPLES_DIR, 'triples_probabilities.csv'), compression = 'gzip')

proba_triples_df.head()

verify_splits(proba_triples_df, None, None,
                  head_col = 'SUBJECT_CUI', rel_col = 'PREDICATE', tail_col = 'OBJECT_CUI',
                  proba_col = 'label_proba')

## Triples Exploration

### COVID19 Risk Factors

SELECTED_KG = 'risk factor covid kg'
save_path = join(SELECTED_TRIPLES, SELECTED_KG)
train, test, val, ent_map, rel_map, ent_type_map = test_read(save_path)

verify_splits(train, test, val, 'h', 'r', 't', 'p')

### COVID19 Time Risk Factors

SELECTED_KG = 'risk_factor_disease_time'
save_path = join(SELECTED_TRIPLES, SELECTED_KG)
raw_train, raw_test, raw_val, train, test, val, ent_map, rel_map, ent_type_map = test_read(save_path)

verify_splits(train, test, val, 'h', 'r', 't', 'p')

## Relational Patterns

save_path = join(SELECTED_TRIPLES, SELECTED_KG)
rel_map = read_json(join(save_path, 'rel_map.json'))

rel_map = {k.lower():v for k, v in rel_map.items()}

def get_relational_patterns(rel_map):
    tran_comp = read_json(join(PROC_UMLS, 'umls_biokg_rel_patterns.json'))
    # transitive relations
    transitive_rels = list(map(rel_map.get, tran_comp['transitive_rels']))
    transitive_rels = [i for i in transitive_rels if i is not None]
    # composition relations
    composition_rels = [list(map(rel_map.get, i)) for i in tran_comp['composition_rels']]
    composition_rels = [i for i in composition_rels if (None not in i) and (len(set(i))==3)]
    save_path = join(SELECTED_TRIPLES, SELECTED_KG, 'relational_patterns.json')
    save_json(save_path, {'composition_rels':composition_rels, 'transitive_rels':transitive_rels})
    return transitive_rels, composition_rels


get_relational_patterns(rel_map)

get_relational_patterns(rel_map)
