==> 5 Format Triples for visualization.ipynb <==
import pandas as pd
from os.path import join
TRIPLES_DIR = 'all_triples'

def get_vis_data():
    df_triples = pd.read_csv(join(TRIPLES_DIR, 'all_data_triples_can.csv'), compression = 'gzip')
    df_ents = pd.read_csv(join(TRIPLES_DIR, 'all_data_entities_can.csv'), compression = 'gzip')
    df_triples = df_triples.rename(columns={'SUBJECT_CUI':'source_id', 'OBJECT_CUI':'target_id'})
    ent_to_type = df_ents.set_index('ENTITY')['SEMTYPE'].to_dict()
    ent_to_name = df_ents.set_index('ENTITY')['NAME'].to_dict()
#     df_triples['edge_id'] = df_triples['PREDICATE']
#     df_triples['edge_type'] = df_triples['PREDICATE']
    df_triples['edge_value'] = 1
    df_triples['source_name'] = df_triples['source_id'].map(ent_to_name)
    df_triples['source_type'] = df_triples['source_id'].map(ent_to_type)
    df_triples['target_name'] = df_triples['target_id'].map(ent_to_name)
    df_triples['target_type'] = df_triples['target_id'].map(ent_to_type)
    df_triples = df_triples.drop(columns=['PREDICATE', 'ORIGIN_ID'])
    return df_triples

vis_triples = get_vis_data()

vis_triples.head()

vis_triples.to_csv(join(TRIPLES_DIR, 'all_data_triples_can_vis.csv'), index = False)
