==> 7-visualization pieces.ipynb <==
import pandas as pd
from glob import glob
from os.path import join
import networkx as nx
from tqdm.notebook import tqdm
import json

PREDICATION_COLS = [
    'PREDICATION_ID', 'SENTENCE_ID', 'PMID', 'PREDICATE', 
    'SUBJECT_CUI', 'SUBJECT_NAME', 'SUBJECT_SEMTYPE', 'SUBJECT_NOVELTY', 'OBJECT_CUI',
    'OBJECT_NAME', 'OBJECT_SEMTYPE', 'OBJECT_NOVELTY', 'Nan1', 'Nan2', 'Nan3']

SENTENCE_DIR = 'SENTENCE'
PREDICATION_DIR = 'PREDICATION'
PREDICATION_AUX_DIR = 'PREDICATION_AUX'


def save_json(save_path, data):
    with open(save_path, 'w') as f:
        json.dump(data, f)
        
def read_json(save_path):
    with open(save_path, 'r') as f:
        data = json.load(f)
    return data

def canonize_cui(cui):
    cacon_form = 'C0000000'
    if len(cui)<8:
        cui = cacon_form[0:-len(cui)]+cui
    elif 'C' not in  cui:
        cui = 'C' + cui
    return cui

def process_mult_cui(cui):
    origin_cui = None
    all_cuis = []
    for cu in cui.split('|'):
        if origin_cui is None:
            cu = canonize_cui(cu)
            origin_cui = cu
        else:
#             cu = origin_cui[0:len(origin_cui)-len(cu)] + cu
            cu = origin_cui[0:-len(cu)] + cu
        all_cuis.append(cu)
    return all_cuis

def process_cui(cui):
    if '|' in cui:
        cuis = process_mult_cui(cui)
    else:
        cuis = [canonize_cui(cui)]
    return cuis

## Example

df = pd.read_csv('Format_sents/form_sent_849503.csv',compression = 'gzip')

df

def vis_formating():

list(df.iloc[0,1:3])

list(df.iloc[2,1:3])

This ... against hospitalizations ... genotype in the same population
This ... against @SUBJECT$ ... genotype @OBJECT$ the same @PREDICAT$

list(df.iloc[3,1:3])

Children with ... rotavirus-confirmed gastroenteritis were individually ...
@PREDICAT$ @OBJECT$ rotavirus-confirmed @SUBJECT$ were individually ...

list(df.iloc[4,1:3])

Nineteen ... human rotavirus vaccine, ....

Nineteen ... @PREDICAT$@OBJECT$@SUBJECT$, ...

list(df.iloc[5,1:3])

Nineteen ... patients had received 2 doses of human rotavirus vaccine ...

Nineteen ... patients had received 2 doses of @PREDICAT$@OBJECT$@SUBJECT$, ...

Nineteen ... @PREDICAT$ had @OBJECT$ 2 doses of human rotavirus @SUBJECT$, ...

list(df.iloc[6,1:3])

list(df.iloc[7,1:3])

all_paths = glob(join('Format_sents', '*'))

all_paths

## Viz interface Data

df = pd.read_csv(join('all_triples', 'triples_probabilities.csv'), compression = 'gzip')

df.head()

FG = nx.from_pandas_edgelist(df, source='SUBJECT_CUI', target='OBJECT_CUI', create_using=nx.DiGraph())

FG

save_json(join('all_triples', 'all_entity_eigenvector_centrality.json'), nx.eigenvector_centrality(FG))

max(list(nx.eigenvector_centrality(FG).values()))

min(list(nx.degree_centrality(FG).values())

max(list(nx.degree_centrality(FG).values()))

## Node Names

all_files = sorted(glob(join(PREDICATION_DIR,'*.gz')))

def canonize_dict(cui_dict):
    fixed_dict = {}
    for cui in cui_dict:
        cuis = process_cui(cui)
        nots = cui_dict[cui].split('|')
        if (len(nots) == 1) & (len(cuis) > 1) :
            nots = nots*len(cuis)
        fixed_dict.update(dict(zip(cuis, nots)))
    return fixed_dict

def get_entity_names_types():
    all_files = sorted(glob(join(PREDICATION_DIR,'*.gz')))
    all_entity_names = {}
    all_entity_types = {}
    for f_name in tqdm(all_files):
        df = pd.read_csv(f_name,  compression= 'gzip', sep = ',', encoding='iso-8859-1', header=None, names = PREDICATION_COLS)
        cui_dict = canonize_dict(dict(zip(list(df['OBJECT_CUI']), list(df['OBJECT_NAME']))))
        all_entity_names.update(cui_dict)
        cui_dict = canonize_dict(dict(zip(list(df['OBJECT_CUI']), list(df['OBJECT_SEMTYPE']))))
        all_entity_types.update(cui_dict)
        cui_dict = canonize_dict(dict(zip(list(df['SUBJECT_CUI']), list(df['SUBJECT_NAME']))))
        all_entity_names.update(cui_dict)
        cui_dict = canonize_dict(dict(zip(list(df['SUBJECT_CUI']), list(df['SUBJECT_SEMTYPE']))))
        all_entity_types.update(cui_dict)
    return all_entity_names, all_entity_types



all_entity_names, all_entity_types = get_entity_names_types()

save_json(join('all_triples', 'all_entity_names.json'), all_entity_names)

save_json(join('all_triples', 'all_entity_types.json'), all_entity_types)

len(all_entity_names)

len(all_entity_types)

df[1]

## Full Data

schema = {
        'source_col' : 'SUBJECT_CUI',
        'rel_col' : 'PREDICATE',
        'dest_col' : 'OBJECT_CUI',
        'source_score' : None,
        'rel_score' : None,
        'dest_score' : None,
        'source_type' : None,
        'rel_type' : None,
        'dest_type' : None,
        'triple_attr_col':None,
        'source_attr_col':None,
        'target_attr_col':None,
    }

def visual_triples():
    df = pd.read_csv(join('all_triples', 'triples_probabilities.csv'), compression = 'gzip')
    all_entity_names = read_json(join('all_triples', 'all_entity_names.json'))
    all_entity_types = read_json(join('all_triples', 'all_entity_types.json'))
    all_entity_ec = read_json(join('all_triples', 'all_entity_eigenvector_centrality.json'))
    df['SUBJECT_EC'] = df['SUBJECT_CUI'].map(all_entity_ec)
    df['OBJECT_EC'] = df['OBJECT_CUI'].map(all_entity_ec)
    df['SUBJECT_NAME'] = df['SUBJECT_CUI'].map(all_entity_names)
    df['OBJECT_NAME'] = df['OBJECT_CUI'].map(all_entity_names)
    df['SUBJECT_TYPE'] = df['SUBJECT_CUI'].map(all_entity_types)
    df['OBJECT_TYPE'] = df['OBJECT_CUI'].map(all_entity_types)
    return df

df = visual_triples()

df

df.to_csv(join('all_triples', 'triples_probabilities_visual.csv'), index = False, compression = 'gzip')
