==> 8- Meta Network exploration.ipynb <==
import pandas as pd
from os.path import join

NET_DIR = '/home/pc/Desktop/AdilStuff/Projects/SemRepMed/umls semantic network/processed_data'

pd.read_csv(join(NET_DIR, 'meta_net.csv'))

pd.read_csv(join(NET_DIR, 'rel_heirarchy.csv'))

df = pd.read_csv(join('all_triples', 'triples_probabilities.csv'), compression = 'gzip')
